Fixed prioritization based on argument matching distance
Currently, we are trying to measure distance between method argument and
argument provided in order to find best fit. Previously unwrapping from
primitive to its counterpart class was assigned 2. But it is creating
issues for situation where 2 primitive type mismatch (each mismatch
means distance 1). So used fraction 1.5. So that it will be bigger
than distance(from Number to Double). But still smaller than other
mistmatch distance.

+++ src/main/java/ca/concordia/jaranalyzer/Models/MethodInfo.java
@@ -37,7 +37,7 @@ public class MethodInfo {
     private String signature;
 
     private int callerClassMatchingDistance;
-    private int argumentMatchingDistance;
+    private double argumentMatchingDistance;
 
     private List<TypeInfo> argumentTypeInfoList;
     private TypeInfo returnTypeInfo;
@@ -340,11 +340,11 @@ public class MethodInfo {
         this.callerClassMatchingDistance = callerClassMatchingDistance;
     }
 
-    public int getArgumentMatchingDistance() {
+    public double getArgumentMatchingDistance() {
         return argumentMatchingDistance;
     }
 
-    public void setArgumentMatchingDistance(int argumentMatchingDistance) {
+    public void setArgumentMatchingDistance(double argumentMatchingDistance) {
         this.argumentMatchingDistance = argumentMatchingDistance;
     }
